* lots of pychecker changes    - remove unreachable code    - strip any N*gios class to nothing    - remove code that could never work    - mark all unused variables    - turn off pychecker when it is most annoying
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@21163 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenModel/ZenStatus.py
@@ -134,7 +134,7 @@ class ZenAvailability(ZenStatus):
         """get current down time in seconds"""
         if self.status < 0: return self.status
         if self.failstart == 0: return 0
-        return round((DateTime() - self.failstart) * 86400)
+        return int(round((DateTime() - self.failstart) * 86400))
         
 
     def getAvailPercent(self, start, end=None):
